+++ test/spec/dashboard.add_product.reducer.ut.js
@@ -3,9 +3,9 @@ import { assign } from 'lodash';
 import { createAction } from 'redux-actions';
 import {
     PRODUCT_SELECTED,
-    WIZARD_COMPLETE,
     GO_TO_STEP,
-    PREVIEW_LOADED
+    PREVIEW_LOADED,
+    COLLECT_PAYMENT
 } from '../../src/actions/product_wizard';
 import { createUuid } from 'rc-uuid';
 import * as TARGETING from '../../src/enums/targeting';
@@ -15,6 +15,7 @@ describe('dashboardAddProductReducer()', function() {
         expect(dashboardAddProductReducer(undefined, 'INIT')).toEqual({
             step: 0,
             previewLoaded: false,
+            checkingIfPaymentRequired: false,
             productData: null,
             targeting: {
                 age: [TARGETING.AGE.ALL],
@@ -32,6 +33,7 @@ describe('dashboardAddProductReducer()', function() {
                 step: 0,
                 productData: null,
                 previewLoaded: false,
+                checkingIfPaymentRequired: false,
                 targeting: {
                     age: [TARGETING.AGE.ALL],
                     gender: TARGETING.GENDER.ALL
@@ -107,56 +109,6 @@ describe('dashboardAddProductReducer()', function() {
             });
         });
 
-        describe(WIZARD_COMPLETE, function() {
-            let productData, targeting;
-
-            beforeEach(function() {
-                productData = {
-                    name: 'The name of my app',
-                    description: 'A description of my app'
-                };
-                targeting = {
-                    age: [TARGETING.AGE.TEENS],
-                    gender: TARGETING.GENDER.FEMALE
-                };
-
-                state.productData = {
-                    name: 'Default name',
-                    description: 'default description',
-                    categories: ['Games', 'Action'],
-                    images: []
-                };
-
-                action = createAction(WIZARD_COMPLETE)({ productData, targeting });
-                newState = dashboardAddProductReducer(state, action);
-            });
-
-            it('should update the productData and targeting and move to step 3', function() {
-                expect(newState).toEqual(assign({}, state, {
-                    productData: assign({}, state.productData, productData),
-                    targeting
-                }));
-            });
-
-            describe('if the targeting options are not defined', function() {
-                beforeEach(function() {
-                    targeting.age = undefined;
-                    targeting.gender = undefined;
-
-                    newState = dashboardAddProductReducer(state, action);
-                });
-
-                it('should use targeting defaults', function() {
-                    expect(newState).toEqual(jasmine.objectContaining({
-                        targeting: {
-                            age: [TARGETING.AGE.ALL],
-                            gender: TARGETING.GENDER.ALL
-                        }
-                    }));
-                });
-            });
-        });
-
         describe(PREVIEW_LOADED, function() {
             beforeEach(function() {
                 this.action = createAction(PREVIEW_LOADED)();
@@ -168,6 +120,50 @@ describe('dashboardAddProductReducer()', function() {
             it('should set previewLoaded to true', function() {
                 expect(this.newState).toEqual(assign({}, state, {
                     previewLoaded: true
+                }));
+            });
+        });
+
+        describe(`${COLLECT_PAYMENT}_PENDING`, function() {
+            beforeEach(function() {
+                this.action = createAction(`${COLLECT_PAYMENT}_PENDING`)();
+
+                this.newState = dashboardAddProductReducer(state, this.action);
+            });
+
+            it('should set checkingIfPaymentRequired to true', function() {
+                expect(this.newState).toEqual(assign({}, state, {
+                    checkingIfPaymentRequired: true
+                }));
+            });
+        });
+
+        describe(`${COLLECT_PAYMENT}_FULFILLED`, function() {
+            beforeEach(function() {
+                this.action = createAction(`${COLLECT_PAYMENT}_FULFILLED`)();
+                state.checkingIfPaymentRequired = true;
+
+                this.newState = dashboardAddProductReducer(state, this.action);
+            });
+
+            it('should set checkingIfPaymentRequired to false', function() {
+                expect(this.newState).toEqual(assign({}, state, {
+                    checkingIfPaymentRequired: false
+                }));
+            });
+        });
+
+        describe(`${COLLECT_PAYMENT}_REJECTED`, function() {
+            beforeEach(function() {
+                this.action = createAction(`${COLLECT_PAYMENT}_REJECTED`)();
+                state.checkingIfPaymentRequired = true;
+
+                this.newState = dashboardAddProductReducer(state, this.action);
+            });
+
+            it('should make checkingIfPaymentRequired to false', function() {
+                expect(this.newState).toEqual(assign({}, state, {
+                    checkingIfPaymentRequired: false
                 }));
             });
         });