+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Showcase
 
+## v0.2.0 (May 27, 2016)
 * *[v0.2.0-rc1]*
   * [FEATURE]: Add a loading animation in the ad preview
   * [FIX]: Make sure the user can see their free trial length after
     signing up
+  * [FIX]: Fix an issue where the name + description of a campaign could
+    end up not being set
+  * [FEATURE]: Add support for free trials that do not require a credit
+    card
+* *[/v0.2.0-rc1]*
 
 ## v0.1.0 (May 25, 2016)
 * *[v0.1.0-rc10]*